Move IdealSizeInvocation to its own file. No functional changes.
